Create wrapping App fro Admin Requests
