+++ app/views/admin/requests/index.html.erb
@@ -1,55 +1,3 @@
-<div class="grid grid--justify-space-around">
-  <div class="grid__col-10 tabs" id="admin-requests-tabs">
-    <ul class="tab-menu">
-      <li class="tab-link">
-        <button
-          role="button"
-          class="tab-button"
-          data-tab-id="pending"
-        >
-          Pending
-        </button>
-      </li>
-
-      <li class="tab-link">
-        <button
-          role="button"
-          class="tab-button"
-          data-tab-id="approved"
-        >
-          Approved
-        </button>
-      </li>
-
-      <li class="tab-link">
-        <button
-          role="button"
-          class="tab-button"
-          data-tab-id="declined"
-        >
-          Declined
-        </button>
-      </li>
-    </ul>
-
-    <div class="content">
-      <div class="tab-content panel" id="pending">
-        <div class="vue-enable-admin-requests">
-          <review-region-requests request-status="pending" />
-        </div>
-      </div>
-
-      <div class="tab-content panel" id="approved">
-        <div class="vue-enable-admin-requests">
-          <review-region-requests request-status="approved" />
-        </div>
-      </div>
-
-      <div class="tab-content panel" id="declined">
-        <div class="vue-enable-admin-requests">
-          <review-region-requests request-status="declined" />
-        </div>
-      </div>
-    </div>
-  </div>
+<div id="vue-enable-admin-requests">
+  <App />
 </div>
