+++ spec/javascript/__mocks__/sweetalert2.js
@@ -1,0 +1,5 @@
+const swal = jest.fn(() => {
+
+})
+
+export default swal
